Initial commit for Giles.
Features to note:

Does rate limiting on outgoing data so we don't swamp
our python based prototype. Sends a new message every
5 milliseconds.

Data send is a monotonically increasing integer.

Verifies that incoming data exists as a message that we
previously sent. It does this as a string comparison that
is case sensitive.