Write collected data to files.
Previously giles did realtime comparison of messages to
and printed "bad message" if the message wasn't one that it
had sent.

Now, we record all sent and received messages to individual
files on program termination (when SIGTERM is received).
These files can then be processed by other tools to verify
the correctness of a given run.