Compare giles output to determine test result.
Previously giles output "bad message" to stdout any time it got a
message that it hadn't previously seen. This was changed. Now,
on receiving SIGTERM giles will output, in order, a file of every
message it sent and other of every message it received, in order
they were sent/received.

For simple, "did we have errors", we can diff the two files. If there
are any differences then something went wrong in terms of:

duplication of messages
dropping of messages
garbling of messages
reordering of messages

From here, we can build tooling as part of giles to check for
any such errors as part of post processing.

This commit is a temporary step to there were we update dagon
to report failures (as it always reported passing after we
removed the output of "bad message" from giles in a previous
commit)