Add periodic statistics collection
Periodic statistics collection is implemented as a
BaseEventLoop.call_later recursive call back.
At each iteration, the throughput and latnecy histograms are popped
from the state object, after which they will be recreated by the next
call to increment stats in the main process.

Currently the statistics are emitted to the global LOGGER object,
but this can easily be changed in the `emit_statistics` function.