Split Giles into two binaries
Nisan, John and I decided that it was time for Giles to be split
into two different binaries. Configuring Giles to be the sender
of multiple input streams forced the issue. We looked at 2 ways
to tackle.

1. Handle in single binary with configuration for each stream
2. Split into sender/receiver and have a sender per stream

Option 1 requires a complicated configuration, probably file based
and if we managed to do via command line, it would be painful to
put the arguments together. By going with split sender/receiver,
we can maintain using simple command line arguments to spin up a
sender and it should allow us when the time comes (if it ever comes)
to scale more effectively. In the end though, that is a secondary
concern. This change is about keeping it easy to work with Giles
via the command line. Given that Giles is primarly controlled via
Dagon, this simplicity makes sense to us.