Allow Giles to shut itself down
Previously, Dagon was responsible for shutting down Giles.
Giles would run until it got a SIGTERM and would then write
out its data and exit. This lead to a scenario where sometimes
Buffy would be running slowly (perhaps because of Spike doing
delays or running on the RPi) and the Giles receiver would be
shut down before all the data arrived.

What we do now is allow Giles to have more control.

Giles sender is passed the number of messages to send. It sends
those messages and then writes out its data and exits.

Giles receiver is passed two different values. The number of seconds
since starting up that it should wait for a message to arrive. If
a message doesn't arrive within that amount of time, the receiver
assumes something has gone wrong and exits. Additionally, you need
to provide the amount of time after any message is received that
you should wait for another to arrive before deciding the run is done.

Via configuration/dagon, these values should be adjusted to reflect
the reality of individual environments. When running on an RPi cluster,
the times to wait should be higher as the processing capability has
been decreased. If the setup time for the Buffy cluster under test is
large, the time to wait for the first message to arrive can be
increased.

Initially we considered giving the Giles receiver knowledge of the
number of messages that would be produced by the sender. We realized
that in the end, this number didn't matter. If we know the sender
will send 10 messages, that tells us nothing about how many might
arrive. Each could get duplicated resulting in 20 total. Each could
be dropped, resulting in 0. The number sent is only loosely connected
to the number received. We therefore decided to not take the number
received into account when deciding when to shut down. We might miss
data with an incorrect shutdown time but given that the Buffy cluster
could delay a message for hours, days etc, we need to make a decision
of when a test is over.

From the Dagon side of things, rather than waiting a fixed amount of
time before shutting down Giles and Buffy, Dagon will now wait for
both the Giles sender and receiver to write out their final data before
shutting down Buffy and concluding a given test run.