Implement metrics output in worker
Use --metrics-address to enable emitting a JSON serialized statistics
collection to a metrics receiver at the given address.
Statistics are emitted at a frequency based on the --stats-period
parameter (default 60 seconds).
The metrics emission is handled as a regular input->func->output
cycle, except the functions are chosen based on input_type
which is given by choose_input() at the beginning of each iteration
of the processing engine's main loop.
If --metrics-address is not provided, no metrics are emitted
(although this does not prevent stats from being emitted to the
log).