Fix latency histogram bin assignment
Rounding log_10 of values < 1 (e.g. negative powers)
resulted in latencies being assigned to the bin below
the one they should be assigned to.
This resulted in incorrect latency histograms, obviously.

"round" was replaced with "math.ceil" to ensure latencies
are assigned to the correct bin (the smallest bin whose value
is greated than the latency value itself).