Enable multi-output for streams
Use multiple --output-address parameters to define a list of output
addresses. This is used to create a list of udp_outputs (replacing
udp_output, and similarly for stats_output).
In the run_engine loop, if the object returned from func(input)
is a tuple, it is assigned to choice and output, otherwise
choice is defined to be 0.
hash(choice) % len(outputs) is then used to select the output
from the output list which is used to handle the output from func().