Add initial support for Wesley
Wesley is the black box test verification system. It compares the
expected received messages from a Buffy topology (based on the sent
messages and a function that generates the expected messages from the
sent messages) with the actual received messages to see if they match,
then compares this match status with the expected match status. If the
expected match status and actual match status are the same then the
test passes, otherwise it fails.

The motivation for Wesley is to have a way to build black box tests
for Buffy that can be run as part of a continuous integration system
to ensure that the overall behavior of Buffy is correct.

Wesley exists as a sort of application framework; tests in Wesley are
standalone executables that use the framework. The test writer is
responsible for implementing classes that:
  * extract data from where it is stored (in this case, in CSV files)
  * represent collections of sent and received messages
  * transform sent messages into the expected received messages

For more information, please see Buffy/wesley/README.md.