Create FlowGraph to represent flows
FlowGraph is a graph with choices built in. It represents
a flow. A FlowGraph can be converted to a set of graphs
via the flowgraphs_to_set() algorithm. In theory we can
check a traversal (a graph without choices) against a flow
by checking whether it is isomorphic to one of the graphs
in the flow set. Graph isomorphism is a hard problem, so
we'll probably need to use a different strategy for real-time
reactions.