Clean up Wesley README.md
Removed the version string which was part of a versioning scheme that
is not longer used.

Some of the terms in the README were difficult to follow because I had
used 1- and 2-word variable names for things (like `S` for sent
messages and `Re` for expected results).

I had used the terms "validator" and "verifier" somewhat
interchangably. I want to be consistent, so I've changed "validate",
"validation", and "validator" to "verify", "verification", and
"verifier".

I stated that the current tests only need a comparison function that
looks for an exact match, but in the future we may need other ways of
comparing results, or we may offer the test implementer a way to
specify a comparison function.