Separate Wesley verifier and verifier-cli
There was a lot of stuff in verfier.pony. Half of it was related to
the actual work of verification, the other half was the part that
handled command line arguments and set up data structures for the
verifier. Based on a suggestion by John, I've separated these pieces
into two separate files, verifier.pony and verifier-cli.pony.