Add wordcount functions
Add the initial code for the Python version of wordcount.

There are actually two different wordcount implementations.

basic_wordcount -- single node that takes input messages that contain
words, splits them, then writes a message per word in the incoming
message with a new count for that word.

wordcount_split and wordcount_count -- wordcount_split splits incoming
messages (which contain words) and sends a message per word to the
next node. wordcount_count takes messages that contain one word and
increments the count for that word. wordcount_split is configured to
act as a router so that it can consistently send words to different
wordcount_count nodes.