Initial pass a "send computation to data"
One of our big ideas with Buffy is to encapsulate state
and send computation to the state rather than the other
way around. The commit is the first step towards introducing
that.

It includes commented out final code in steps.pony that requires
ponyc issue #807 to be fixed. Once that done, we will have a
usable, generic first basic pass. In the meantime, a simple
version of an app that uses this was added "state-avg-of-avgs".

Once a couple other changes are done, the marketspread app will
use the functionality and be the best example of how it works.