Fix Spike bug where junk data was sent after a dropped connection.
When there is a pending read on the TCPConnection and Spike drops the
connection, it was reading data from the buffer but not sending it
along to the wrapped Notify. However, the wrapped Notify is still
expecting it and if we don't send it along, then we'll end up sending
junk to match its expectations, leading to a segfault when combined
with deserialize.