Fix machine word boxing problem in Buffy jr
The major source of Buffy jr slowness was the
boxing of machine words in our FirstPass "take"
method.

We took in an any and then match as a U64. This
is fine unless we want to send that U64 to another
actor whose type is Any or some union type and we
need to box this new U64 in order to send over.

A brand new U64 would have the same problem. IE.

_next.take(a, b, c, U64(1))

would have the same problem. That U64 needs to be boxed.

This is a huge problem in Buffy. Luckily we can solve this
with.... PARAMETERIZATION!

This commit parameterizes the "take" method to avoid boxing.

Without this change, throughput on my machine in in the 65k
to 90k per second range. After this change, I am getting
155k to 175k per second. If I turn off metrics that we know
is a performance issue and can be improved, I am consistently
getting throughputs in the 185-190k per second range.