Dagon arbitrary arguments (#232)
* add state to dagon process manager

* add external start of giles senders msg

* add dagon-notifier for sending messages to dagon

* add delay senders arg and response when senders started

* add GilesSendersStarted msg

* change if/elseif to match

* Add parsing of wrapper to time commands with Dagon

We need a way to wrap a command started by Dagon inside a gnu_time
command or similar to get executions statistics for each run of a
given command. We add it for forked processes first.

* Add support for wrapping commands that run inside a container

* Add unlimited wrapper_args

* Add TLS deactivation note. Patch Giles-Receiver Docker image.

For now we need to build Giles-Receiver Docker images on top of
"busybox" to have the "time" utility available for our example
wrapper config file.