Introduce "state computations" to marketspread jr
A small step towards "Buffy-ness".
Run our computations against our NBBO data actor
as code defined outside of it. That is, arbitrary
computation.

This change resulted in no throughput or memory
usage changes on my laptop. It was run with
the same setup as you see in the package
docstring.