Add "source" concept to market spread jr
This moves all our logic out of the TCPConnection notifiers.
You now supply a "source" class that handles processing incoming
data and routing it to its first location.

There was no throughput or memory change on my laptop with this
change. Details on how to run are in the docstring.

Next step... to generify our usage of TCPConnectionNotifiers?
Maybe... that might be part of topology construction that I am
pushing off til the end.