Add tcp out output to market-spread-jr
No perf or memory impact. Every tenth order for a symbol,
it sends that symbol as output. This doesn't really push
cross actor GC so the next step is to send variable GC'able
data and see the impact.

After that, computations currently don't return output which
makes them not very real world useful. Will be correcting that
soon.