More gc realistic market-spread-jr output
We know "reject" one out of ten orders for a symbol
(this didn't change), however rather than sending
the same value (the symbol) out via tcp on rejections,
we send the count of orders we've seen for this symbol
as a string.

This should exercise cross actor gc more.

No performance difference. I saw a bit more memory
usage that looked like it might want to climb
but I did multiple runs (5) and it never got past
45 megs which it had settled on after 3 runs.

Looking good.