Add pipeline metrics to marketspread jr
Will need to be updated for when computations return real results.
I tried using a class for Source name and Source timestamp and
the Order stream started queuing like crazy. NBBO ran fine but
it was getting all the scheduler time compared to the Order
stream. Actually, it might not have been scheduler time, it
could be something about ASIO. Either way, it wasn't good.
Definitely something to discuss with Sylvan.