Make `name()` methods return const char* in C++ API
Prior to this commit, `name()` methods returned char *, which meant
that one could not return a string literal. This meant that the
programmer had to jump through hoops to return a new array, which
wasn't really the right thing to do anyway because of the way the
string was being used in Pony.