Set outgoing route_id in Router after sending message downstream
We need to keep track of the outgoing route_id for bookkeeping
(resilience) purposes. To do this we hand down a ref of the
CreditFlowProducer which now has a update_route_id method. The
Router then calls update_route_id once it has sent the message
downstream. If we have to queue the message we also enqueue the
ref to the CreditFlowProducer.