Fix memory leak in Queue
Queue was never truncating the internal data structure used to
store data. The `_data` array would grow and grow and grow never
getting smaller. It was a sad queue with bad self esteem for it
knew that it was causing poor performance as it hogged memory.

Now, its a happy queue. :) goes the queue.

QUEUE QUEUE!