Remove guard-rail in EventLog
The "ifdef" around pushing to a buffer was an "extra" protection.
If you managed to execute that code while not running resilience,
there would be a large memory leak. However, that code should
never be executed when resilience isnt on. It even being executed
would be an error. I removed the guard rail so if that code
is incorrectly executed, we'll have an obvious sign of it- the
massive memory leak.