Remove unneeded Producer match statement from routers
Previously when wiring things together, we passed producer
parameter as Producer | None. The None has been removed for
a while but we still had the match statements hanging around.

This commits removes the unneeded match statements.
Removes the TODO's to do this
And otherwise updates variable names etc to match