Address large performance issue lurking in Wallaroo
While testing watermark batching code on a branch, we changed one small
bit of code and introduced a large performance regression. The change
was to set an array used for tracking to the watermark batch size. In
this particular case, 10_000. This change should have been a
performance improvement, but, it wasn't. Things got a good bit worse.
After a good deal of testing and poking, we figured out why.

Our problem was that by increasing the initial memory used by our Step,
after GCing, there was a baseline of memory that would always remain
used. This changed the value at which we would next gc. Before the
addition of the array with 10_000 elements, we were, without realizing
it, gc'ing our steps after each batch of messages handled. After the
change, that was no longer the case.

We have since determined that based on how Wallaroo works, that any
Step or Step like thing (TCPSink, Boundary) should be gc'ing at a
regular interval. If they aren't then the Step/Step like thing that
sent them messages is stuck holding on to large amount of data and
poor performance ensues.

This changes introduces an intentional garbage collection run after
each time a Step handles a batch of messages. It is quite likely
that TCPSink and Boundary need this as well but those need to be
tested and added after testing.