Use library for converting bytes to numbers in arizona-jr
Prior to this commit we were using hand-rolled functions for
converting between bytes (in messages coming over the network) and
numbers. We now use libraries that provide the functions.

Doing this in a protable way is actually surprisingly difficult. This
gist shows one way to do it:

  https://gist.github.com/panzi/6856583

After looking around a bit, it appears that JetBrains uses almost
exactly the same technique in their Thrift code, so if this is the
best they can do then I'll accept it and move on.