Handle becoming muted while handling data in TCPSource
Prior to this change, if a TCPSource read 100 bytes of
data and became muted after processing 50 bytes of that
data, it would contine to process the next 50 bytes.
In a backpressure scenario, this could cause an overload.
With this change, after each call to our notify's received
method, we check to see if we have become muted and if we
have, we stop processing at that point leaving the remaining
data in our expect buffer.

In the case of not running with expect, if we read 100 bytes
and become muted while handling it, we will check to see if
muted and exit early rather than going through the usual
"read_again" logic.