Make reading via TCPSource re-entrant safe
There's a nasty possible bug in TCPSource before applying this change.

Because `_mute` and `_unmute` are synchronous methods on TCPSource and
because they can be called inside `_pending_reads` via code executing
in `notify.received`, its possible to mute and unmute a source
while executing `_pending_reads`.

Prior to this change, if you were to unmute from inside a pending
reads via the above scenario, `_unmute` would synchronously call
`_pending_reads` to re-enter that function and start processing
data out of order. Further, we could end up in an infinite loop
so long as enough data existed.

The logic in `_pending_reads` needs to be extracted and cleaned up
in general. This commit doesn't do that. It only addresses the
re-entrance problem.