Fail when TCPSourceRouteCallbackHandler receives incorrect producer type
Because route callback handlers are generic, you can receive any type
of class that implements `Producer` in your various
`RouteCallbackHandler` methods.

In the case of `TCPSourceRouteCallbackHandler`, the only valid type
is a `TCPSource`. This commit updates the handler to fail the
running wallaroo application if it receives any other type.

Failure is the appropriate response here as getting a type other than
`TCPSource` is indicative of a wiring failure.