Stop using ManagedObject in the C++ API
Prior to this commit all C++ API objects held a ManagedObject that was
responsible for handling memory management of the underlying C++
object. Allocating this extra object will have a performance impact,
so it is hoped that performance improves by removing it.

The Pony objects are now directly responsible for memory management of
their underlying C++ objects.

ManagedObject was also responsible for serializing the C++ object, so
this responsiblity has also been moved to the higher-level objects as
well.