Set maximum number of credits to ever give out on a route
If we only have 1 producer, we don't want to allow it to
send an insane number of messges. This is applicable to
TCPSink and Boundary at the moment. Really the issue
is that they have way too many distributable credits
as a fan in source, however, we don't have a good way to
deal with that. This is a short term fix.