Fail application startup if outgoing boundary can't connect
This can currently silently hide errors that would result
in us chasing ghosts. This isn't the correct way to handle.
The correct way involves having an actor that takes us
through an application startup lifecycle and fails if
an outgoing boundary can't connect. We don't however, have
that yet so in the meantime, calling `Fail()` is a good,
simple blunt tool to wield.