Resilience without collapsing
Should be run with backpressure on.

-D backpressure -D resilience

This can do NBBO at a batch size of 100 from Giles sender without
failing over and decent latency. However, it doesn't currently
sync to disk (only flush) and as such, might lose data in a crash.

I've found 2 keys to getting this to perform well. Don't queue
anything. Try not to sync. The two are in conflict so this is
going to be a delicate balancing act to get it working fully.