Improve stability and performance when reading from TCP
After a good deal of testing, we found that Wallaroo's
various TCP reading implementation's performed poorly
when reading from a socket that was experiencing TCP
back pressure; that is, that the OS buffer always
had data in it.

After a lot of testing, we found the values for max_size
and for the newly introduced max_read that are hard
coded in by this commit produce very good performance
results with market spread and also great stability.

Whereas under back pressure previously we had awful
performance and stability, in the same situations
with this change (and the new adapative batch sizes
for actor scheduling in our ponyc) that we get much
better stability.