Fix memory leak in arizona source app
If an order had the same order id as a previously recorded order then
a newly created order object replaced the old one without deleting the
old one.

This is happening because we sometimes test by looping over a file, so
orders may be repeated before they are filled or canceled.

For now we are looking into the order map to see if an order exists,
and if it does then we don't add the new one.