Fix memory explosion in Giles receiver
There's a bug in Pony where if you pass an iso object in a message send
and its received as a val, then we get a massive memory explosion.
There's an open Pony issue for this.

This commit fixes that leak in Giles receiver so that we don't get
unnatural memory growth.