Turn on acking and queuing across boundary.
First step towards running with resilience in multiworker
mode again. As part of this, for stability, I switched
from using our queue to an array to keep the upstream backup
in the boundary. When using the queue, if we got momentarily
overloaded, it would grow to a large size and then remain at
that size causing gc related issues as it held on to a lot
more objects that were created at the source. The array we
are now using doesn't have any noticeable impact on performance.
We'd expect it to as it is allocating memory over and over.

Possible reasons for lack of degradation in performance:

Cost of holding items for longer in the queue than array equals
the cost of allocations for array

The additional cost of the array memory allocations is lost in
the noise of sending over the network