Upgrade Giles Sender to work with lastest master
Previously, we were stuck working off of 14.0.5 as the
output characteristics of the sender changed when we switched
the Pony implementation of writev from doing the syscall `write` over
and over to using the actual `writev` syscall.

With this change, Giles sender now compiles with master and replicates
the old Pony TCPConnection `writev` behavior by calling `write` in
a loop. This isn't the best for sender performance but it gives us
a consistent performance and we don't have to tweak any existing
Giles sender usage instructions to match a different output profile.