Reformat commands in ARIZONA_UBUNTU.md to make them more readable
There are a number of long commands in the instructions for running
Arizona. They are hard to read on github because you have to scroll
through the text area to get to the end.

This change breaks the commands up across several lines so that they
should disply without scrolling in github.