Make multiple-worker resilience recovery work: Step 1
So, this commit has a number of changes required to get multiple-
worker recovery to work. Changes range from startup/initialization
on recovery to reconnecting an outgoing boundary to bug fixes
for decoding DataMsg inside of a ReplayMsg and a fix in replay_route
in the router. There's probably some other stuff I'm forgetting.