Fix alfred o_seq_id/i_seq_id issue
Prior to this commit alfie was receiving `o_seq_id` for watermarks
and `i_seq_id` for normal log entries which meant that on recovery
most of the recovered log entries would be thrown away. This change
makes it so that alfie is always given `o_seq_id` and as a result
most of the log entries are recovered successfully and very few are
thrown away.