Revert to using "_final" in CPPData
As a workaround to pony_alloc_final being slow, we had inserted a workaround
that handled the CPP-allocated memory manually, given that we knew the exact
point where we would no longer need the objects created in CPP-land. These were
marked by a comment "SUPER-EVIL-DANGER-ZONE", just to remind everyone that we
were playing with fire. This commit removes those sections and restores the use
of pony_alloc_final now that pony_alloc_final is no longer slow.