Remove Spike Delay
When revisiting the Spike Delay functionality, I realized it
was a poor means of trying to simulate network delays. It doesn't
really simulate those delays. What it is good at doing is simulating
hidden buffering inside an application that is invisible and we
can't react to. Basically, its not a good Spike.