Rename lib/* pony files and directories to match Pony file naming
This is step 1 in the great renaming.

Any files in lib/ that had `-` in the name was changed to `_`.
This leads to a number of import changes that need to be
made throughout the respository.

Additionally renaming of directories and files outside of `lib`
are needed as well. Those will come in another commit.

File naming might not match with Pony naming standards in terms
of content. This will come in another commit.