Enable nominal join of a running cluster
If a worker is run with the --join/-j parameter and passed the control channel
address of any worker in an existing cluster, then it will nominally join the
cluster, which at the moment means that it will register with the monitoring hub
but is otherwise unknown to the cluster and no part of the topology exists on
it. The contacted worker replies with metrics related info and the new worker
contacts the monitoring hub using that information.