Add support for building an entire Wallaroo application in C++
This commit makes it possible to create a Wallaroo application using
on C++ code. The biggest addition is new classes that support
application creation and pipeline setup.

This also includes some API cleanup.

Get rid of EncodableData class

It doesn't seem to make sense to force the application developer to
subclass a special class of data in order to encode it. If the
application developer wishes to add methods to data to make it easier
to encode it, then they can, but this should not be required.

Requiring data that goes to the encoder to have a special type also
meant that these data classes were implemented differently than
standard data classes, which meant that more work would be required
when changing a pipeline layout. If data that had been going to a
computation now needed to go to a sink, it's type would have to be
changed.

Get rid of unused size arguments in serialization and encoding

The serialization and encoding methods and functions took size
arguments, but these arguments were not being set because it wasn't
possible to determine the sizes in the code. Therefore, these
arguments are no longer passed to these functions.

Change C++ StateBuilder class use use `build()` method

The method name now matches the intention of this class, which is to
be a builder for State objects.

Add ComputationBuilder to C++ API

The C++ API now uses a `ComputationBuilder` class rather than a
builder function to create a computation. This is more consistent with
other types which also use builder classes rather than builder
functions..

Fix bad return types introduced earlier in c++ API