Add RouterRegistry and, on migration, update routers and routes
The RouterRegistry keeps track of all routers that might
be updated across a worker, as well as all the steps that
will need to be informed about such updates. Hooks allow
routers to be updated at the registry on state migration
and a new control message is used to inform the rest of
the workers of a new step that must be added to their
routers.