Add partition rebalancing algo for new worker join
RouterRegistry has a hook called migrate_partition_steps()
that can be called when a new worker joins and we are
ready to rebalance a partition. It kicks off an algo
that determines how to approximate rebalancing steps in
the partition such that there are an equal number on
each worker. When steps are successfully migrated,
messages are sent between workers to notify them that
they should update their routers accordingly.