Allow simple increase in giles send rate
This will be used for grow to fit demo. IT SHOUD NOT BE CONSIDERED
"THIS IS GOOD" CODE. At this point, Giles has had tons of code
added on and is getting long in the tooth. This works for our
demo needs:

After X number of messages, increase batch size by Y

If you wanted to do "variable rate" sending, this is absolutely
not how you would do it. Giles needs love. But this will do for
now. Yes, that is right, this is a hate commit.

Sorry.

Not sorry.