Temporary fix for AWS us-east-1 vpc issues
Without this fix the aws us-east-1 vpc subnets get reordered
and terraform tries to delete and recreate them preventing clusters
from being created.

This is just a temporary/quick fix. A more proper fix would make
the subnets for each vpc stable so it wouldn't change if a new
AZ is added.