Add Makefiles for C++ sample applications
These Makefiles are intended to serve as an example of how to build a
C++ Wallaroo application.

The files are named `Makefile.example` rather than `Makefile` in order
to avoid causing problems with the recursive Makefiles that are part
of the project. As such, the C++ application will not be built by the
recursive Makefiles and consequently will not be built as part of CI.

Makefiles that work with the recursive Makefiles and CI will be added
later.