Allow for randomly varying Giles sender batch size
Changes -j option to be amount to vary batch size by.
If not passed, then there is no variation. If passed,
then on any given run, the batch size will be increased
between 0 and -j for that given batch run.

This commit removes the previous temporary "one time spike"
functionality.