Make length-framing optional at Decoder
A user can determine whether wallaroo uses the length-framing of
`<header-length><header><body>` stream reading scheme via their
FramedSourceNotifier's `header_length()` function.

If `header_length() => 0` (also now the default at the interface),
then the `<body>` scheme is used, where the entire buffer is passed
directly to the FramedSourceHandler's `decode()` function.

If `header_length()` returns a value >0, then that number
of bytes are first read and passed to `payload_length` to determine
the correct length of bytes to read and pass to the `decode`
function.