Use correct giles sender has_next check in binary multi-file sources
Prior to this commit, in the MultiFileBinaryDataSource and the
MultiFileVariableBinaryDataSource would only check if the current
source has a next message to send or not. This would incorrectly
result in an error thrown if the current source ran out of messages
to send but there was another source (or a repeat of the current
source) from which messages were supposed to be read and sent.

This commit fixes the two sources to do a `has_next` check for
at the multi-file source level instead of just the current source.

Resolves #598