Move giles receiver expected and metrics logic to coordinators
Prior to this commit, giles receiver was tracking metrics and
number of messages received so far on a per connection basis.
This would lead to incorrect behavior for expected when the
number expected was received across multiple connections instead
of a single connection.

This commit moves the expect related logic into the coordinator
so that number of messages can be tracked across all connections.
It also adds the expect logic into the Store so that it can
discard any additional messages that might be received instead
of potentially writing partial or invalid data. The commit also
moves the metrics logic into the coordinator to give a holistic
picture of messages received including throughput across all
connections.

Resolves #222