Add a size argument to SourceDecoder::decode
The size argument indicates the number of bytes in the array that is
passed to the decoder. This makes it possible to write logic that
prevents reading beyond the end of the buffer, and also allows the
developer to use this length to read variable-length messages where
the overall message length can be used to determine the lengths of
parts of the message (for example strings or arrays of uniformly sized
items.

The counter-app and market-spread-cpp app have been updated to use the
new API.