Use strings instead of bytearrays in Python encoders and decoders
Change the Python API to use strings instead of byte arrays as the
input to decoders and the output from encoders. Strings and bytearrays
are almost the same, except that bytearrays are mutable. The bytes
going into and coming out of the encoders and decoders should not be
mutable, so strings should be used here.

There are changes to:
* the C and Pony code that implements the Python API
* the documentation
* the existing example programs

Fixes #711.