Add support for multi-worker apps and resilience in Python API
Multi-worker applications and resilience rely on Pony's serialisation
system. We added custom Pony serialisation support to the Pony objects
that wrap the Python objects.

In order to support serialization for either multi-worker
applications, resilience, or both, the application developer must
provide the following two Python functions:
* `serialize(obj)`
* `deserialize(bs)`

The `serialize(...)` function takes an object to be serialized and
returns a string representation of that object. A simple
implementation would do something like call `pickle.dumps(obj)` and
return the result.

The `deserialize(...)` function takes a string representing an object,
and returns the object itself. If, as above, `pickle.dumps(obj)` was
used to create the string then this function could call
`pickle.loads(bs)` to obtain the object, and then return the resulting
object.

Unlike the C++ API, the layers above these functions take care of
encoding and decoding a length that is stored along with the
serialized representation. Therefore the user does not have to worry
about including this information.

If either of these functions is not provided then inter-worker
marshaling and resilience will not work. This is fine as long as the
application is always run on a single worker and resilience is not
enabled.

In order to support resilience the API for stateful computation
changed; the user is now required to return a two-tuple that consists
of the outgoing message (or None for no outgoing message) and either
True (to indicate that the state should be persisted after this
computation for resilience) or False (to indicate that the state
should not be persisted after this computation). This give the user
control over when the state is persisted, so that if there is not a
state change then the application can avoid the performance penalty of
persisting the state.

Note: Pony's custom serialization system is implemented in such a way
that it first gets the number of bytes needed to serialize an object,
then later passes a c byte buffer of that size to the function that
actually encodes the object. In order to simplify the Python API, we
have opted to call the `serialize(...)` function and get the length of
the returned string when the size is requested, and then call
`serialize(...)` a second time when the actual serialized
representation is requested. This means that `serialize(...)` is
called twice. This decision was made because we felt that most users
would use a 3rd part serialization system (like pickle) to generate
the serialized representation of objects, and these packages might not
provide a means for determining the size of the serialized
representation without going through the process of serializing. There
are performance optimizations that can be implemented to minimize the
impact of calling this function twice, such as caching the serialized
representation when the first call is made. If needed, we may add a
function that can be called explicitly to get the size of the
serialized representation.